avance

+++ Backend/servicio-venta/src/main/java/com/example/servicioventa/service/VentaService.java
@@ -1,6 +1,8 @@
 package com.example.servicioventa.service;
 
+
 import com.example.servicioventa.entity.Venta;
+import jakarta.transaction.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -9,7 +11,6 @@ public interface VentaService {
     public List<Venta> listar();
     public Venta guardarVenta(Venta venta);
     public Venta actualizar(Venta venta);
-    public Optional<Venta> listarPorId(Integer id);
-    public void eliminarPorId(Integer id);
-
+    public Optional<Venta> listarPorId(Long id);
+    public void eliminarPorId(Long id);
 }
